// src/views/verse/lib/cannon/cannon.d.ts
  export class AABB {
    lowerBound: Vec3
    upperBound: Vec3

    constructor(options?: IAABBOptions)

    clone(): AABB
    copy(aabb: AABB): void
    extend(aabb: AABB): void
    getCorners(a: Vec3, b: Vec3, c: Vec3, d: Vec3, e: Vec3, f: Vec3, g: Vec3, h: Vec3): void
    overlaps(aabb: AABB): boolean
    setFromPoints(points: Vec3[], position?: Vec3, quaternion?: Quaternion, skinSize?: number): AABB
    toLocalFrame(frame: Transform, target: AABB): AABB
    toWorldFrame(frame: Transform, target: AABB): AABB
  }

  export class ArrayCollisionMatrix {
    matrix: Mat3[]

    get(i: number, j: number): number
    set(i: number, j: number, value: number): void
    reset(): void
    setNumObjects(n: number): void
  }

  export class BroadPhase {
    world: World
    useBoundingBoxes: boolean
    dirty: boolean

    collisionPairs(world: World, p1: Body[], p2: Body[]): void
    needBroadphaseCollision(bodyA: Body, bodyB: Body): boolean
    intersectionTest(bodyA: Body, bodyB: Body, pairs1: Body[], pairs2: Body[]): void
    doBoundingSphereBroadphase(bodyA: Body, bodyB: Body, pairs1: Body[], pairs2: Body[]): void
    doBoundingBoxBroadphase(bodyA: Body, bodyB: Body, pairs1: Body[], pairs2: Body[]): void
    makePairsUnique(pairs1: Body[], pairs2: Body[]): void
    setWorld(world: World): void
    boundingSphereCheck(bodyA: Body, bodyB: Body): boolean
    aabbQuery(world: World, aabb: AABB, result: Body[]): Body[]
  }

  export class GridBroadphase extends BroadPhase {
    nx: number
    ny: number
    nz: number
    aabbMin: Vec3
    aabbMax: Vec3
    bins: any[]

    constructor(aabbMin?: Vec3, aabbMax?: Vec3, nx?: number, ny?: number, nz?: number)
  }

  export class NaiveBroadphase extends BroadPhase {}
// ...
  export class RaycastResult {
    rayFromWorld: Vec3
    rayToWorld: Vec3
    hitNormalWorld: Vec3
    hitPointWorld: Vec3
    hasHit: boolean
    shape: Shape
    body: Body
    distance: number

    reset(): void
    set(
      rayFromWorld: Vec3,
      rayToWorld: Vec3,
      hitNormalWorld: Vec3,
      hitPointWorld: Vec3,
      shape: Shape,
      body: Body,
      distance: number
    ): void
  }

  export class SAPBroadphase extends BroadPhase {
    static insertionSortX(a: any[]): any[]
    static insertionSortY(a: any[]): any[]
    static insertionSortZ(a: any[]): any[]
    static checkBounds(bi: Body, bj: Body, axisIndex?: number): boolean

    axisList: any[]
    world: World
    axisIndex: number

    constructor(world?: World)

    autoDetectAxis(): void
    aabbQuery(world: World, aabb: AABB, result?: Body[]): Body[]
  }

  export interface IConstraintOptions {
    collideConnected?: boolean
    wakeUpBodies?: boolean
  }

  export class Constraint {
    equations: any[]
    bodyA: Body
    bodyB: Body
    id: number
    collideConnected: boolean

    constructor(bodyA: Body, bodyB: Body, options?: IConstraintOptions)

    update(): void
    disable(): void
    enable(): void
  }
// ...
  export class FrictionEquation extends Equation {
    constructor(bi: Body, bj: Body, slipForce: number)
  }
// ...
  export class ContactEquation extends Equation {
    restitution: number
    ri: Vec3
    rj: Vec3
    penetrationVec: Vec3
    ni: Vec3
    rixn: Vec3
    rjxn: Vec3
    invIi: Mat3
    invIj: Mat3
    biInvInertiaTimesRixn: Vec3
    bjInvInertiaTimesRjxn: Vec3

    constructor(bi: Body, bj: Body)
  }

  export interface IContactMaterialOptions {
    friction?: number
    restitution?: number
    contactEquationStiffness?: number
    contactEquationRelaxation?: number
    frictionEquationStiffness?: number
    frictionEquationRelaxation?: number
  }

  export class ContactMaterial {
    id: number
    materials: Material[]
    friction: number
    restitution: number
    contactEquationStiffness: number
    contactEquationRelaxation: number
    frictionEquationStiffness: number
    frictionEquationRelaxation: number

    constructor(m1: Material, m2: Material, options?: IContactMaterialOptions)
  }

  export class Material {
    name: string
    id: number
    friction: number
    restitution: number

    constructor(name: string)
  }

  export class JacobianElement {
    spatial: Vec3
    rotational: Vec3

    multiplyElement(element: JacobianElement): number
    multiplyVectors(spacial: Vec3, rotational: Vec3): number
  }

  export class Mat3 {
    constructor(elements?: number[])

    identity(): void
    setZero(): void
    setTrace(vec3: Vec3): void
    getTrace(target: Vec3): void
    vmult(v: Vec3, target?: Vec3): Vec3
    smult(s: number): void
    mmult(m: Mat3): Mat3
    scale(v: Vec3, target?: Mat3): Mat3
    solve(b: Vec3, target?: Vec3): Vec3
    e(row: number, column: number, value?: number): number
    copy(source: Mat3): Mat3
    toString(): string
    reverse(target?: Mat3): Mat3
    setRotationFromQuaternion(q: Quaternion): Mat3
    transpose(target?: Mat3): Mat3
  }

  export class Quaternion {
    x: number
    y: number
    z: number
    w: number

    constructor(x?: number, y?: number, z?: number, w?: number)

    set(x: number, y: number, z: number, w: number): void
    toString(): string
    toArray(): number[]
    setFromAxisAngle(axis: Vec3, angle: number): void
    toAxisAngle(targetAxis?: Vec3): any[]
    setFromVectors(u: Vec3, v: Vec3): void
    mult(q: Quaternion, target?: Quaternion): Quaternion
    inverse(target?: Quaternion): Quaternion
    conjugate(target?: Quaternion): Quaternion
    normalize(): void
    normalizeFast(): void
    vmult(v: Vec3, target?: Vec3): Vec3
    copy(source: Quaternion): Quaternion
    toEuler(target: Vec3, order?: string): void
    setFromEuler(x: number, y: number, z: number, order?: string): Quaternion
    clone(): Quaternion
  }
// ...
  export class Vec3 {
    static ZERO: Vec3

    constructor(x?: number, y?: number, z?: number)

    x: number
    y: number
    z: number

    cross(v: Vec3, target: Vec3): Vec3
    set(x: number, y: number, z: number): Vec3
    setZero(): void
    vadd(v: Vec3, target: Vec3): Vec3
    vsub(v: Vec3, target: Vec3): Vec3
    crossmat(): Mat3
    normalize(): number
    unit(target: Vec3): Vec3
    norm(): number

    /**
     * Get the length of the vector
     * @method length
     * @return {Number}
     */
    length: any
    norm2(): number

    /**
     * Get the squared length of the vector.
     * @method lengthSquared
     * @return {Number}
     */
    lengthSquared: any
    distanceTo(p: Vec3): number
    distanceSquared(p: Vec3): number
    mult(scalar: number, target: Vec3): Vec3
    vmul(vector: number, target: Vec3): Vec3

    /**
     * Multiply the vector with a scalar.
     * @method scale
     * @param {Number} scalar
     * @param {Vec3} target
     * @return {Vec3}
     */
    scale: any
    addScaledVector(scalar: number, vector: Vec3, target: Vec3): Vec3
    dot(v: Vec3): number
    isZero(): boolean
    negate(target: Vec3): Vec3
    tangents(t1: any, t2: any): void
    toString(): string
    toArray(): number[]
    copy(source: Vec3): Vec3
    lerp(v: Vec3, t: number, target: Vec3): void
    almostEquals(v: Vec3, precision: number): boolean
    almostZero(precision: number): boolean
    isAntiparallelTo(v: Vec3, precision: number): boolean
    clone(): Vec3
  }
// ...
  export class Body extends EventTarget {
    static DYNAMIC: number
    static STATIC: number
    static KINEMATIC: number
    static AWAKE: number
    static SLEEPY: number
    static SLEEPING: number
    static sleepyEvent: IEvent
    static sleepEvent: IEvent

    id: number
    world: World
    preStep: Function
    postStep: Function
    vlambda: Vec3
    collisionFilterGroup: number
    collisionFilterMask: number
    collisionResponse: boolean
    position: Vec3
    previousPosition: Vec3
    initPosition: Vec3
    velocity: Vec3
    initVelocity: Vec3
    force: Vec3
    mass: number
    invMass: number
    material: Material
    linearDamping: number
    type: number
    allowSleep: boolean
    sleepState: number
    sleepSpeedLimit: number
    sleepTimeLimit: number
    timeLastSleepy: number
    torque: Vec3
    quaternion: Quaternion
    initQuaternion: Quaternion
    angularVelocity: Vec3
    initAngularVelocity: Vec3
    interpolatedPosition: Vec3
    interpolatedQuaternion: Quaternion
    shapes: Shape[]
    shapeOffsets: any[]
    shapeOrientations: any[]
    inertia: Vec3
    invInertia: Vec3
    invInertiaWorld: Mat3
    invMassSolve: number
    invInertiaSolve: Vec3
    invInteriaWorldSolve: Mat3
    fixedRotation: boolean
    angularDamping: number
    aabb: AABB
    aabbNeedsUpdate: boolean
    wlambda: Vec3

    constructor(options?: IBodyOptions)

    wakeUp(): void
    sleep(): void
    sleepTick(time: number): void
    updateSolveMassProperties(): void
    pointToLocalFrame(worldPoint: Vec3, result?: Vec3): Vec3
    pointToWorldFrame(localPoint: Vec3, result?: Vec3): Vec3
    vectorToWorldFrame(localVector: Vec3, result?: Vec3): Vec3
    addShape(shape: Shape, offset?: Vec3, orientation?: Quaternion): void
    updateBoundingRadius(): void
    computeAABB(): void
    updateInertiaWorld(force: Vec3): void
    applyForce(force: Vec3, worldPoint: Vec3): void
    applyImpulse(impulse: Vec3, worldPoint: Vec3): void
    applyLocalForce(force: Vec3, localPoint: Vec3): void
    applyLocalImpulse(impulse: Vec3, localPoint: Vec3): void
    updateMassProperties(): void
    getVelocityAtWorldPoint(worldPoint: Vec3, result: Vec3): Vec3
  }
// ...
  export class Solver {
    iterations: number
    equations: Equation[]

    solve(dy: number, world: World): number
    addEquation(eq: Equation): void
    removeEquation(eq: Equation): void
    removeAllEquations(): void
  }

  export class SplitSolver extends Solver {
    subsolver: Solver

    constructor(subsolver: Solver)

    solve(dy: number, world: World): number
  }

  export class EventTarget {
    addEventListener(type: string, listener: Function): EventTarget
    hasEventListener(type: string, listener: Function): boolean
    removeEventListener(type: string, listener: Function): EventTarget
    dispatchEvent(event: IEvent): IEvent
  }

  export class Pool {
    objects: any[]
    type: any[]

    release(): any
    get(): any
    constructObject(): any
  }

  export class TupleDictionary {
    data: {
      keys: any[]
    }

    get(i: number, j: number): number
    set(i: number, j: number, value: number): void
    reset(): void
  }

  export class Utils {
    static defaults(options?: any, defaults?: any): any
  }

  export class Vec3Pool extends Pool {
    type: any

    constructObject(): Vec3
  }

  export class NarrowPhase {
    contactPointPool: Pool[]
    v3pool: Vec3Pool
  }

  export class World extends EventTarget {
    iterations: number
    dt: number
    allowSleep: boolean
    contacts: ContactEquation[]
    frictionEquations: FrictionEquation[]
    quatNormalizeSkip: number
    quatNormalizeFast: boolean
    time: number
    stepnumber: number
    default_dt: number
    nextId: number
    gravity: Vec3
    broadphase: NaiveBroadphase
    bodies: Body[]
    solver: Solver
    constraints: Constraint[]
    narrowPhase: NarrowPhase
    collisionMatrix: ArrayCollisionMatrix
    collisionMatrixPrevious: ArrayCollisionMatrix
    materials: Material[]
    contactmaterials: ContactMaterial[]
    contactMaterialTable: TupleDictionary
    defaultMaterial: Material
    defaultContactMaterial: ContactMaterial
    doProfiling: boolean
    profile: {
      solve: number
      makeContactConstraints: number
      broadphaser: number
      integrate: number
      narrowphase: number
    }
    subsystems: any[]
    addBodyEvent: IBodyEvent
    removeBodyEvent: IBodyEvent

    getContactMaterial(m1: Material, m2: Material): ContactMaterial
    numObjects(): number
    collisionMatrixTick(): void
    addBody(body: Body): void
    addConstraint(c: any): void
    removeConstraint(c: any): void
    rayTest(from: Vec3, to: Vec3, result: RaycastResult): void
    raycastAll(
      from: Vec3,
      to: Vec3,
      options: {
        collisionFilterMask?: number
        collisionFilterGroup?: number
        skipBackfaces?: boolean
        checkCollisionResponse?: boolean
      },
      callback: Function
    ): boolean
    raycastAny(
      from: Vec3,
      to: Vec3,
      options: {
        collisionFilterMask?: number
        collisionFilterGroup?: number
        skipBackfaces?: boolean
        checkCollisionResponse?: boolean
      },
      result: RaycastResult
    ): boolean
    raycastClosest(
      from: Vec3,
      to: Vec3,
      options: {
        collisionFilterMask?: number
        collisionFilterGroup?: number
        skipBackfaces?: boolean
        checkCollisionResponse?: boolean
      },
      result: RaycastResult
    ): boolean
    remove(body: Body): void
    getBodyById(id: any): any
    getShapeById(id: any): any
    addMaterial(m: Material): void
    addContactMaterial(cmat: ContactMaterial): void
    step(dt: number, timeSinceLastCalled?: number, maxSubSteps?: number): void
    internalStep(dt: any): void
    emitContactEvents: () => void
    clearForces(): void
  }

  export interface IEvent {
    type: string
  }

  export interface IBodyEvent extends IEvent {
    body: Body
  }